fine_tuning/requirements.txt
# Fine-tuning dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
peft>=0.4.0
bitsandbytes>=0.41.0
trl>=0.7.0
wandb>=0.15.0
tensorboard>=2.13.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0
jsonlines>=3.1.0
sentencepiece>=0.1.99
protobuf>=3.20.0

